requirements.txt
numpy~=1.25.0
pandas~=1.5.3
matplotlib~=3.7.2
seaborn
opencv-python
scikit-image~=0.19.3
tqdm~=4.65.0
python-decouple
pylatexenc~=2.10
BeautifulSoup4
pillow~=9.4.0
torchvision~=0.15.2
torcheval~=0.0.6
streamlit~=1.27.1